+++ src/server.ts
@@ -86,7 +86,7 @@ export class ServerManager {
 
       let sbm = window.setStatusBarMessage(`Starting the P5 server at ${root}`);
 
-      this.server = new Server({ root });
+      this.server = new Server({ root, theme: 'grid' });
       await this.server.start();
       const consolePane = new ScriptConsole();
       consolePane.subscribe(this.server);