+++ CHANGELOG.md
@@ -7,6 +7,10 @@ Changed:
 - Reloading a sketch clears the console. A new configuration "P5-server >
   Console: Clear on Reload" restores the old behavior.
 
+Improved:
+
+- Performance improvements
+
 ## [1.1.8] - 2021-10-28
 
 Added: